Network.AWS.WorkSpaces

Description

Amazon WorkSpaces offers you an easy way to provide a cloud-based desktop experience to your end-users. You simply select from a choice of bundles that offer a range of different amounts of CPU, memory, storage, and a choice of applications. Users can connect from a PC, Mac desktop computer, iPad, Kindle, or Android tablet.

Documentation

module Network.AWS.WorkSpaces.CreateWorkspaces

module Network.AWS.WorkSpaces.DescribeWorkspaceBundles

module Network.AWS.WorkSpaces.DescribeWorkspaceDirectories

module Network.AWS.WorkSpaces.DescribeWorkspaces

module Network.AWS.WorkSpaces.RebootWorkspaces

module Network.AWS.WorkSpaces.RebuildWorkspaces

module Network.AWS.WorkSpaces.TerminateWorkspaces

module Network.AWS.WorkSpaces.Types