Network.AWS.WorkSpaces.CreateWorkspaces

Description

Creates one or more WorkSpaces.

This operation is asynchronous and returns before the WorkSpaces are created.

http://docs.aws.amazon.com/workspaces/latest/devguide/API_CreateWorkspaces.html

Request

data CreateWorkspaces

Request constructor

createWorkspaces

Arguments

:: NonEmpty WorkspaceRequest
cwWorkspaces

-> CreateWorkspaces

CreateWorkspaces constructor.

The fields accessible through corresponding lenses are:

cwWorkspaces :: NonEmpty WorkspaceRequest

Request lenses

cwWorkspaces :: Lens' CreateWorkspaces (NonEmpty WorkspaceRequest)

An array of structures that specify the WorkSpaces to create.

Response

data CreateWorkspacesResponse

Response constructor

createWorkspacesResponse :: CreateWorkspacesResponse

CreateWorkspacesResponse constructor.

The fields accessible through corresponding lenses are:

cwrFailedRequests :: [FailedCreateWorkspaceRequest]

cwrPendingRequests :: [Workspace]

Response lenses

cwrFailedRequests :: Lens' CreateWorkspacesResponse [FailedCreateWorkspaceRequest]

An array of structures that represent the WorkSpaces that could not be created.

cwrPendingRequests :: Lens' CreateWorkspacesResponse [Workspace]

An array of structures that represent the WorkSpaces that were created.

Because this operation is asynchronous, the identifier in WorkspaceId is not immediately available. If you immediately call DescribeWorkspaces with this identifier, no information will be returned.