Network.AWS.WorkSpaces.DescribeWorkspaceBundles

Description

Obtains information about the WorkSpace bundles that are available to your account in the specified region.

You can filter the results with either the BundleIds parameter, or the Owner parameter, but not both.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

http://docs.aws.amazon.com/workspaces/latest/devguide/API_DescribeWorkspaceBundles.html

Request

data DescribeWorkspaceBundles

Request constructor

describeWorkspaceBundles

Arguments

:: NonEmpty Text
dwbBundleIds

-> DescribeWorkspaceBundles

DescribeWorkspaceBundles constructor.

The fields accessible through corresponding lenses are:

dwbBundleIds :: NonEmpty Text

dwbNextToken :: Maybe Text

dwbOwner :: Maybe Text

Request lenses

dwbBundleIds :: Lens' DescribeWorkspaceBundles (NonEmpty Text)

An array of strings that contains the identifiers of the bundles to retrieve. This parameter cannot be combined with any other filter parameter.

dwbNextToken :: Lens' DescribeWorkspaceBundles (Maybe Text)

The NextToken value from a previous call to this operation. Pass null if this is the first call.

dwbOwner :: Lens' DescribeWorkspaceBundles (Maybe Text)

The owner of the bundles to retrieve. This parameter cannot be combined with any other filter parameter.

This contains one of the following values:

null - Retrieves the bundles that belong to the account making the call. AMAZON - Retrieves the bundles that are provided by AWS.

Response

data DescribeWorkspaceBundlesResponse

Response constructor

describeWorkspaceBundlesResponse :: DescribeWorkspaceBundlesResponse

DescribeWorkspaceBundlesResponse constructor.

The fields accessible through corresponding lenses are:

dwbrBundles :: [WorkspaceBundle]

dwbrNextToken :: Maybe Text

Response lenses

dwbrBundles :: Lens' DescribeWorkspaceBundlesResponse [WorkspaceBundle]

An array of structures that contain information about the bundles.

dwbrNextToken :: Lens' DescribeWorkspaceBundlesResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to this operation to retrieve the next set of items. This token is valid for one day and must be used within that timeframe.