Network.AWS.WorkSpaces.DescribeWorkspaceDirectories

Description

Retrieves information about the AWS Directory Service directories in the region that are registered with Amazon WorkSpaces and are available to your account.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

http://docs.aws.amazon.com/workspaces/latest/devguide/API_DescribeWorkspaceDirectories.html

Request

data DescribeWorkspaceDirectories

Request constructor

describeWorkspaceDirectories

Arguments

:: NonEmpty Text
dwdDirectoryIds

-> DescribeWorkspaceDirectories

DescribeWorkspaceDirectories constructor.

The fields accessible through corresponding lenses are:

dwdDirectoryIds :: NonEmpty Text

dwdNextToken :: Maybe Text

Request lenses

dwdDirectoryIds :: Lens' DescribeWorkspaceDirectories (NonEmpty Text)

An array of strings that contains the directory identifiers to retrieve information for. If this member is null, all directories are retrieved.

dwdNextToken :: Lens' DescribeWorkspaceDirectories (Maybe Text)

The NextToken value from a previous call to this operation. Pass null if this is the first call.

Response

data DescribeWorkspaceDirectoriesResponse

Response constructor

describeWorkspaceDirectoriesResponse :: DescribeWorkspaceDirectoriesResponse

DescribeWorkspaceDirectoriesResponse constructor.

The fields accessible through corresponding lenses are:

dwdrDirectories :: [WorkspaceDirectory]

dwdrNextToken :: Maybe Text

Response lenses

dwdrDirectories :: Lens' DescribeWorkspaceDirectoriesResponse [WorkspaceDirectory]

An array of structures that contain information about the directories.

dwdrNextToken :: Lens' DescribeWorkspaceDirectoriesResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to this operation to retrieve the next set of items. This token is valid for one day and must be used within that timeframe.