Network.AWS.WorkSpaces.DescribeWorkspaces

Description

Obtains information about the specified WorkSpaces.

Only one of the filter parameters, such as BundleId, DirectoryId, or WorkspaceIds, can be specified at a time.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

http://docs.aws.amazon.com/workspaces/latest/devguide/API_DescribeWorkspaces.html

Request

data DescribeWorkspaces

Request constructor

describeWorkspaces

Arguments

:: NonEmpty Text
dwWorkspaceIds

-> DescribeWorkspaces

DescribeWorkspaces constructor.

The fields accessible through corresponding lenses are:

dwBundleId :: Maybe Text

dwDirectoryId :: Maybe Text

dwLimit :: Maybe Natural

dwNextToken :: Maybe Text

dwUserName :: Maybe Text

dwWorkspaceIds :: NonEmpty Text

Request lenses

dwBundleId :: Lens' DescribeWorkspaces (Maybe Text)

The identifier of a bundle to obtain the WorkSpaces for. All WorkSpaces that are created from this bundle will be retrieved. This parameter cannot be combined with any other filter parameter.

dwDirectoryId :: Lens' DescribeWorkspaces (Maybe Text)

Specifies the directory identifier to which to limit the WorkSpaces. Optionally, you can specify a specific directory user with the UserName parameter. This parameter cannot be combined with any other filter parameter.

dwLimit :: Lens' DescribeWorkspaces (Maybe Natural)

The maximum number of items to return.

dwNextToken :: Lens' DescribeWorkspaces (Maybe Text)

The NextToken value from a previous call to this operation. Pass null if this is the first call.

dwUserName :: Lens' DescribeWorkspaces (Maybe Text)

Used with the DirectoryId parameter to specify the directory user for which to obtain the WorkSpace.

dwWorkspaceIds :: Lens' DescribeWorkspaces (NonEmpty Text)

An array of strings that contain the identifiers of the WorkSpaces for which to retrieve information. This parameter cannot be combined with any other filter parameter.

Because the CreateWorkspaces operation is asynchronous, the identifier returned by CreateWorkspaces is not immediately available. If you immediately call DescribeWorkspaces with this identifier, no information will be returned.

Response

data DescribeWorkspacesResponse

Response constructor

describeWorkspacesResponse :: DescribeWorkspacesResponse

DescribeWorkspacesResponse constructor.

The fields accessible through corresponding lenses are:

dwrNextToken :: Maybe Text

dwrWorkspaces :: [Workspace]

Response lenses

dwrNextToken :: Lens' DescribeWorkspacesResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to this operation to retrieve the next set of items. This token is valid for one day and must be used within that timeframe.

dwrWorkspaces :: Lens' DescribeWorkspacesResponse [Workspace]

An array of structures that contain the information about the WorkSpaces.

Because the CreateWorkspaces operation is asynchronous, some of this information may be incomplete for a newly-created WorkSpace.