Network.AWS.WorkSpaces.DescribeWorkspacesConnectionStatus

Description

Describes the connection status of a specified WorkSpace.

Creating a Request

describeWorkspacesConnectionStatus :: DescribeWorkspacesConnectionStatus

Creates a value of DescribeWorkspacesConnectionStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwcsWorkspaceIds - An array of strings that contain the identifiers of the WorkSpaces.

dwcsNextToken - The next token of the request.

data DescribeWorkspacesConnectionStatus

See: describeWorkspacesConnectionStatus smart constructor.

Request Lenses

dwcsWorkspaceIds :: Lens' DescribeWorkspacesConnectionStatus (Maybe (NonEmpty Text))

An array of strings that contain the identifiers of the WorkSpaces.

dwcsNextToken :: Lens' DescribeWorkspacesConnectionStatus (Maybe Text)

The next token of the request.

Destructuring the Response

describeWorkspacesConnectionStatusResponse

Arguments

:: Int
dwcsrsResponseStatus

-> DescribeWorkspacesConnectionStatusResponse

Creates a value of DescribeWorkspacesConnectionStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwcsrsNextToken - The next token of the result.

dwcsrsWorkspacesConnectionStatus - The connection status of the WorkSpace.

dwcsrsResponseStatus - -- | The response status code.

data DescribeWorkspacesConnectionStatusResponse

See: describeWorkspacesConnectionStatusResponse smart constructor.

Response Lenses

dwcsrsNextToken :: Lens' DescribeWorkspacesConnectionStatusResponse (Maybe Text)

The next token of the result.

dwcsrsWorkspacesConnectionStatus :: Lens' DescribeWorkspacesConnectionStatusResponse [WorkspaceConnectionStatus]

The connection status of the WorkSpace.

dwcsrsResponseStatus :: Lens' DescribeWorkspacesConnectionStatusResponse Int

- | The response status code.