Network.AWS.WorkSpaces.RebuildWorkspaces

Description

Rebuilds the specified WorkSpaces.

Rebuilding a WorkSpace is a potentially destructive action that can result in the loss of data. Rebuilding a WorkSpace causes the following to occur:

The system is restored to the image of the bundle that the WorkSpace is created from. Any applications that have been installed, or system settings that have been made since the WorkSpace was created will be lost. The data drive (D drive) is re-created from the last automatic snapshot taken of the data drive. The current contents of the data drive are overwritten. Automatic snapshots of the data drive are taken every 12 hours, so the snapshot can be as much as 12 hours old. To be able to rebuild a WorkSpace, the WorkSpace must have a State of AVAILABLE or ERROR.

This operation is asynchronous and will return before the WorkSpaces have been completely rebuilt.

http://docs.aws.amazon.com/workspaces/latest/devguide/API_RebuildWorkspaces.html

Request

data RebuildWorkspaces

Request constructor

rebuildWorkspaces

Arguments

:: NonEmpty RebuildRequest
rwRebuildWorkspaceRequests

-> RebuildWorkspaces

RebuildWorkspaces constructor.

The fields accessible through corresponding lenses are:

rwRebuildWorkspaceRequests :: NonEmpty RebuildRequest

Request lenses

rwRebuildWorkspaceRequests :: Lens' RebuildWorkspaces (NonEmpty RebuildRequest)

An array of structures that specify the WorkSpaces to rebuild.

Response

data RebuildWorkspacesResponse

Response constructor

rebuildWorkspacesResponse :: RebuildWorkspacesResponse

RebuildWorkspacesResponse constructor.

The fields accessible through corresponding lenses are:

rwrFailedRequests :: [FailedWorkspaceChangeRequest]

Response lenses

rwrFailedRequests :: Lens' RebuildWorkspacesResponse [FailedWorkspaceChangeRequest]

An array of structures that represent any WorkSpaces that could not be rebuilt.