Network.AWS.WorkSpaces.StopWorkspaces

Description

Stops the specified WorkSpaces. The WorkSpaces must have a running mode of AutoStop and a state of AVAILABLE, IMPAIRED, UNHEALTHY, or ERROR.

Creating a Request

stopWorkspaces

Arguments

:: NonEmpty StopRequest
swStopWorkspaceRequests

-> StopWorkspaces

Creates a value of StopWorkspaces with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

swStopWorkspaceRequests - The requests.

data StopWorkspaces

See: stopWorkspaces smart constructor.

Request Lenses

swStopWorkspaceRequests :: Lens' StopWorkspaces (NonEmpty StopRequest)

The requests.

Destructuring the Response

stopWorkspacesResponse

Arguments

:: Int
srsResponseStatus

-> StopWorkspacesResponse

Creates a value of StopWorkspacesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsFailedRequests - The failed requests.

srsResponseStatus - -- | The response status code.

data StopWorkspacesResponse

See: stopWorkspacesResponse smart constructor.

Response Lenses

srsFailedRequests :: Lens' StopWorkspacesResponse [FailedWorkspaceChangeRequest]

The failed requests.

srsResponseStatus :: Lens' StopWorkspacesResponse Int

- | The response status code.