Network.AWS.WorkSpaces.TerminateWorkspaces

Description

Terminates the specified WorkSpaces.

Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is not maintained and will be destroyed. If you need to archive any user data, contact Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except SUSPENDED.

This operation is asynchronous and will return before the WorkSpaces have been completely terminated.

http://docs.aws.amazon.com/workspaces/latest/devguide/API_TerminateWorkspaces.html

Request

data TerminateWorkspaces

Request constructor

terminateWorkspaces

Arguments

:: NonEmpty TerminateRequest
twTerminateWorkspaceRequests

-> TerminateWorkspaces

TerminateWorkspaces constructor.

The fields accessible through corresponding lenses are:

twTerminateWorkspaceRequests :: NonEmpty TerminateRequest

Request lenses

twTerminateWorkspaceRequests :: Lens' TerminateWorkspaces (NonEmpty TerminateRequest)

An array of structures that specify the WorkSpaces to terminate.

Response

data TerminateWorkspacesResponse

Response constructor

terminateWorkspacesResponse :: TerminateWorkspacesResponse

TerminateWorkspacesResponse constructor.

The fields accessible through corresponding lenses are:

twrFailedRequests :: [FailedWorkspaceChangeRequest]

Response lenses

twrFailedRequests :: Lens' TerminateWorkspacesResponse [FailedWorkspaceChangeRequest]

An array of structures that represent any WorkSpaces that could not be terminated.