Network.AWS.WorkSpaces.Types

Service

data WorkSpaces

Version 2015-04-08 of the Amazon WorkSpaces service.

Error

data JSONError :: *

WorkspaceRequest

data WorkspaceRequest

workspaceRequest

Arguments

:: Text
wDirectoryId

-> Text
wUserName

-> Text
wBundleId

-> WorkspaceRequest

WorkspaceRequest constructor.

The fields accessible through corresponding lenses are:

wBundleId :: Text

wDirectoryId :: Text

wUserName :: Text

wBundleId :: Lens' WorkspaceRequest Text

The identifier of the bundle to create the WorkSpace from. You can use the DescribeWorkspaceBundles operation to obtain a list of the bundles that are available.

wDirectoryId :: Lens' WorkspaceRequest Text

The identifier of the AWS Directory Service directory to create the WorkSpace in. You can use the DescribeWorkspaceDirectories operation to obtain a list of the directories that are available.

wUserName :: Lens' WorkspaceRequest Text

The username that the WorkSpace is assigned to. This username must exist in the AWS Directory Service directory specified by the DirectoryId member.

WorkspaceDirectory

data WorkspaceDirectory

workspaceDirectory :: WorkspaceDirectory

WorkspaceDirectory constructor.

The fields accessible through corresponding lenses are:

wdAlias :: Maybe Text

wdCustomerUserName :: Maybe Text

wdDirectoryId :: Maybe Text

wdDirectoryName :: Maybe Text

wdDirectoryType :: Maybe WorkspaceDirectoryType

wdDnsIpAddresses :: [Text]

wdIamRoleId :: Maybe Text

wdRegistrationCode :: Maybe Text

wdState :: Maybe WorkspaceDirectoryState

wdSubnetIds :: [Text]

wdWorkspaceCreationProperties :: Maybe DefaultWorkspaceCreationProperties

wdWorkspaceSecurityGroupId :: Maybe Text

wdAlias :: Lens' WorkspaceDirectory (Maybe Text)

The directory alias.

wdCustomerUserName :: Lens' WorkspaceDirectory (Maybe Text)

The user name for the service account.

wdDirectoryId :: Lens' WorkspaceDirectory (Maybe Text)

The directory identifier.

wdDirectoryName :: Lens' WorkspaceDirectory (Maybe Text)

The name of the directory.

wdDirectoryType :: Lens' WorkspaceDirectory (Maybe WorkspaceDirectoryType)

The directory type.

wdDnsIpAddresses :: Lens' WorkspaceDirectory [Text]

An array of strings that contains the IP addresses of the DNS servers for the directory.

wdIamRoleId :: Lens' WorkspaceDirectory (Maybe Text)

The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

wdRegistrationCode :: Lens' WorkspaceDirectory (Maybe Text)

The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

wdState :: Lens' WorkspaceDirectory (Maybe WorkspaceDirectoryState)

The state of the directory's registration with Amazon WorkSpaces

wdSubnetIds :: Lens' WorkspaceDirectory [Text]

An array of strings that contains the identifiers of the subnets used with the directory.

wdWorkspaceCreationProperties :: Lens' WorkspaceDirectory (Maybe DefaultWorkspaceCreationProperties)

A structure that specifies the default creation properties for all WorkSpaces in the directory.

wdWorkspaceSecurityGroupId :: Lens' WorkspaceDirectory (Maybe Text)

The identifier of the security group that is assigned to new WorkSpaces.

Compute

data Compute

Constructors

Performance
PERFORMANCE

Standard
STANDARD

Value
VALUE

Workspace

data Workspace

workspace :: Workspace

Workspace constructor.

The fields accessible through corresponding lenses are:

w1BundleId :: Maybe Text

w1DirectoryId :: Maybe Text

w1ErrorCode :: Maybe Text

w1ErrorMessage :: Maybe Text

w1IpAddress :: Maybe Text

w1State :: Maybe WorkspaceState

w1SubnetId :: Maybe Text

w1UserName :: Maybe Text

w1WorkspaceId :: Maybe Text

w1BundleId :: Lens' Workspace (Maybe Text)

The identifier of the bundle that the WorkSpace was created from.

w1DirectoryId :: Lens' Workspace (Maybe Text)

The identifier of the AWS Directory Service directory that the WorkSpace belongs to.

w1ErrorCode :: Lens' Workspace (Maybe Text)

If the WorkSpace could not be created, this contains the error code.

w1ErrorMessage :: Lens' Workspace (Maybe Text)

If the WorkSpace could not be created, this contains a textual error message that describes the failure.

w1IpAddress :: Lens' Workspace (Maybe Text)

The IP address of the WorkSpace.

w1State :: Lens' Workspace (Maybe WorkspaceState)

The operational state of the WorkSpace.

w1SubnetId :: Lens' Workspace (Maybe Text)

The identifier of the subnet that the WorkSpace is in.

w1UserName :: Lens' Workspace (Maybe Text)

The user that the WorkSpace is assigned to.

w1WorkspaceId :: Lens' Workspace (Maybe Text)

The identifier of the WorkSpace.

RebuildRequest

data RebuildRequest

rebuildRequest

Arguments

:: Text
rWorkspaceId

-> RebuildRequest

RebuildRequest constructor.

The fields accessible through corresponding lenses are:

rWorkspaceId :: Text

rWorkspaceId :: Lens' RebuildRequest Text

The identifier of the WorkSpace to rebuild.

FailedCreateWorkspaceRequest

data FailedCreateWorkspaceRequest

failedCreateWorkspaceRequest :: FailedCreateWorkspaceRequest

FailedCreateWorkspaceRequest constructor.

The fields accessible through corresponding lenses are:

fcwErrorCode :: Maybe Text

fcwErrorMessage :: Maybe Text

fcwWorkspaceRequest :: Maybe WorkspaceRequest

fcwErrorCode :: Lens' FailedCreateWorkspaceRequest (Maybe Text)

The error code.

fcwErrorMessage :: Lens' FailedCreateWorkspaceRequest (Maybe Text)

The textual error message.

fcwWorkspaceRequest :: Lens' FailedCreateWorkspaceRequest (Maybe WorkspaceRequest)

A WorkspaceRequest object that contains the information about the WorkSpace that could not be created.

WorkspaceDirectoryState

data WorkspaceDirectoryState

Constructors

Deregistered
DEREGISTERED

Deregistering
DEREGISTERING

Error
ERROR

Registered
REGISTERED

Registering
REGISTERING

FailedWorkspaceChangeRequest

data FailedWorkspaceChangeRequest

failedWorkspaceChangeRequest :: FailedWorkspaceChangeRequest

FailedWorkspaceChangeRequest constructor.

The fields accessible through corresponding lenses are:

fwcErrorCode :: Maybe Text

fwcErrorMessage :: Maybe Text

fwcWorkspaceId :: Maybe Text

fwcErrorCode :: Lens' FailedWorkspaceChangeRequest (Maybe Text)

The error code.

fwcErrorMessage :: Lens' FailedWorkspaceChangeRequest (Maybe Text)

The textual error message.

fwcWorkspaceId :: Lens' FailedWorkspaceChangeRequest (Maybe Text)

The identifier of the WorkSpace.

WorkspaceDirectoryType

data WorkspaceDirectoryType

Constructors

AdConnector
AD_CONNECTOR

SimpleAd
SIMPLE_AD

WorkspaceState

data WorkspaceState

Constructors

WSAvailable
AVAILABLE

WSError
ERROR

WSImpaired
IMPAIRED

WSPending
PENDING

WSRebooting
REBOOTING

WSRebuilding
REBUILDING

WSSuspended
SUSPENDED

WSTerminated
TERMINATED

WSTerminating
TERMINATING

WSUnhealthy
UNHEALTHY

ComputeType

data ComputeType

computeType :: ComputeType

ComputeType constructor.

The fields accessible through corresponding lenses are:

ctName :: Maybe Compute

ctName :: Lens' ComputeType (Maybe Compute)

The name of the compute type for the bundle.

RebootRequest

data RebootRequest

rebootRequest

Arguments

:: Text
r1WorkspaceId

-> RebootRequest

RebootRequest constructor.

The fields accessible through corresponding lenses are:

r1WorkspaceId :: Text

r1WorkspaceId :: Lens' RebootRequest Text

The identifier of the WorkSpace to reboot.

UserStorage

data UserStorage

userStorage :: UserStorage

UserStorage constructor.

The fields accessible through corresponding lenses are:

usCapacity :: Maybe Text

usCapacity :: Lens' UserStorage (Maybe Text)

The amount of user storage for the bundle.

TerminateRequest

data TerminateRequest

terminateRequest

Arguments

:: Text
tWorkspaceId

-> TerminateRequest

TerminateRequest constructor.

The fields accessible through corresponding lenses are:

tWorkspaceId :: Text

tWorkspaceId :: Lens' TerminateRequest Text

The identifier of the WorkSpace to terminate.

DefaultWorkspaceCreationProperties

data DefaultWorkspaceCreationProperties

defaultWorkspaceCreationProperties :: DefaultWorkspaceCreationProperties

DefaultWorkspaceCreationProperties constructor.

The fields accessible through corresponding lenses are:

dwcpCustomSecurityGroupId :: Maybe Text

dwcpDefaultOu :: Maybe Text

dwcpEnableInternetAccess :: Maybe Bool

dwcpEnableWorkDocs :: Maybe Bool

dwcpUserEnabledAsLocalAdministrator :: Maybe Bool

dwcpCustomSecurityGroupId :: Lens' DefaultWorkspaceCreationProperties (Maybe Text)

The identifier of any custom security groups that are applied to the WorkSpaces when they are created.

dwcpDefaultOu :: Lens' DefaultWorkspaceCreationProperties (Maybe Text)

The organizational unit (OU) in the directory that the WorkSpace machine accounts are placed in.

dwcpEnableInternetAccess :: Lens' DefaultWorkspaceCreationProperties (Maybe Bool)

A public IP address will be attached to all WorkSpaces that are created or rebuilt.

dwcpEnableWorkDocs :: Lens' DefaultWorkspaceCreationProperties (Maybe Bool)

Specifies if the directory is enabled for Amazon WorkDocs.

dwcpUserEnabledAsLocalAdministrator :: Lens' DefaultWorkspaceCreationProperties (Maybe Bool)

The WorkSpace user is an administrator on the WorkSpace.

WorkspaceBundle

data WorkspaceBundle

workspaceBundle :: WorkspaceBundle

WorkspaceBundle constructor.

The fields accessible through corresponding lenses are:

wbBundleId :: Maybe Text

wbComputeType :: Maybe ComputeType

wbDescription :: Maybe Text

wbName :: Maybe Text

wbOwner :: Maybe Text

wbUserStorage :: Maybe UserStorage

wbBundleId :: Lens' WorkspaceBundle (Maybe Text)

The bundle identifier.

wbComputeType :: Lens' WorkspaceBundle (Maybe ComputeType)

A ComputeType object that specifies the compute type for the bundle.

wbDescription :: Lens' WorkspaceBundle (Maybe Text)

The bundle description.

wbName :: Lens' WorkspaceBundle (Maybe Text)

The name of the bundle.

wbOwner :: Lens' WorkspaceBundle (Maybe Text)

The owner of the bundle. This contains the owner's account identifier, or AMAZON if the bundle is provided by AWS.

wbUserStorage :: Lens' WorkspaceBundle (Maybe UserStorage)

A UserStorage object that specifies the amount of user storage that the bundle contains.