Network.AWS.WorkSpaces

Description

Amazon WorkSpaces Service

This reference provides detailed information about the Amazon WorkSpaces operations.

Service Configuration

workSpaces :: Service

API version 2015-04-08 of the Amazon WorkSpaces SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by WorkSpaces.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The user is not authorized to access a resource.

ResourceUnavailableException

_ResourceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is not available.

InvalidParameterValuesException

_InvalidParameterValuesException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameter values are not valid.

OperationInProgressException

_OperationInProgressException :: AsError a => Getting (First ServiceError) a ServiceError

The properties of this WorkSpace are currently being modified. Try again in a moment.

ResourceLimitExceededException

_ResourceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your resource limits have been exceeded.

InvalidResourceStateException

_InvalidResourceStateException :: AsError a => Getting (First ServiceError) a ServiceError

The state of the WorkSpace is not valid for this operation.

UnsupportedWorkspaceConfigurationException

_UnsupportedWorkspaceConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

The configuration of this WorkSpace is not supported for this operation. For more information, see the Amazon WorkSpaces Administration Guide .

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource could not be found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ModifyWorkspaceProperties

module Network.AWS.WorkSpaces.ModifyWorkspaceProperties

DescribeTags

module Network.AWS.WorkSpaces.DescribeTags

DescribeWorkspaceDirectories (Paginated)

module Network.AWS.WorkSpaces.DescribeWorkspaceDirectories

DescribeWorkspaceBundles (Paginated)

module Network.AWS.WorkSpaces.DescribeWorkspaceBundles

RebuildWorkspaces

module Network.AWS.WorkSpaces.RebuildWorkspaces

CreateTags

module Network.AWS.WorkSpaces.CreateTags

DeleteTags

module Network.AWS.WorkSpaces.DeleteTags

StopWorkspaces

module Network.AWS.WorkSpaces.StopWorkspaces

DescribeWorkspacesConnectionStatus

module Network.AWS.WorkSpaces.DescribeWorkspacesConnectionStatus

RebootWorkspaces

module Network.AWS.WorkSpaces.RebootWorkspaces

TerminateWorkspaces

module Network.AWS.WorkSpaces.TerminateWorkspaces

CreateWorkspaces

module Network.AWS.WorkSpaces.CreateWorkspaces

DescribeWorkspaces (Paginated)

module Network.AWS.WorkSpaces.DescribeWorkspaces

StartWorkspaces

module Network.AWS.WorkSpaces.StartWorkspaces

Types

Compute

data Compute

Constructors

Performance

Standard

Value

ConnectionState

data ConnectionState

Constructors

Connected

Disconnected

Unknown

RunningMode

data RunningMode

Constructors

AlwaysOn

AutoStop

WorkspaceDirectoryState

data WorkspaceDirectoryState

Constructors

Deregistered

Deregistering

Error'

Registered

Registering

WorkspaceDirectoryType

data WorkspaceDirectoryType

Constructors

AdConnector

SimpleAd

WorkspaceState

data WorkspaceState

Constructors

WSAvailable

WSError'

WSImpaired

WSMaintenance

WSPending

WSRebooting

WSRebuilding

WSStarting

WSStopped

WSStopping

WSSuspended

WSTerminated

WSTerminating

WSUnhealthy

ComputeType

data ComputeType

Contains information about the compute type of a WorkSpace bundle.

See: computeType smart constructor.

computeType :: ComputeType

Creates a value of ComputeType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctName - The name of the compute type for the bundle.

ctName :: Lens' ComputeType (Maybe Compute)

The name of the compute type for the bundle.

DefaultWorkspaceCreationProperties

data DefaultWorkspaceCreationProperties

Contains default WorkSpace creation information.

See: defaultWorkspaceCreationProperties smart constructor.

defaultWorkspaceCreationProperties :: DefaultWorkspaceCreationProperties

Creates a value of DefaultWorkspaceCreationProperties with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwcpCustomSecurityGroupId - The identifier of any custom security groups that are applied to the WorkSpaces when they are created.

dwcpUserEnabledAsLocalAdministrator - The WorkSpace user is an administrator on the WorkSpace.

dwcpEnableWorkDocs - Specifies if the directory is enabled for Amazon WorkDocs.

dwcpEnableInternetAccess - A public IP address will be attached to all WorkSpaces that are created or rebuilt.

dwcpDefaultOu - The organizational unit (OU) in the directory that the WorkSpace machine accounts are placed in.

dwcpCustomSecurityGroupId :: Lens' DefaultWorkspaceCreationProperties (Maybe Text)

The identifier of any custom security groups that are applied to the WorkSpaces when they are created.

dwcpUserEnabledAsLocalAdministrator :: Lens' DefaultWorkspaceCreationProperties (Maybe Bool)

The WorkSpace user is an administrator on the WorkSpace.

dwcpEnableWorkDocs :: Lens' DefaultWorkspaceCreationProperties (Maybe Bool)

Specifies if the directory is enabled for Amazon WorkDocs.

dwcpEnableInternetAccess :: Lens' DefaultWorkspaceCreationProperties (Maybe Bool)

A public IP address will be attached to all WorkSpaces that are created or rebuilt.

dwcpDefaultOu :: Lens' DefaultWorkspaceCreationProperties (Maybe Text)

The organizational unit (OU) in the directory that the WorkSpace machine accounts are placed in.

FailedCreateWorkspaceRequest

data FailedCreateWorkspaceRequest

Contains information about a WorkSpace that could not be created.

See: failedCreateWorkspaceRequest smart constructor.

failedCreateWorkspaceRequest :: FailedCreateWorkspaceRequest

Creates a value of FailedCreateWorkspaceRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fcwrWorkspaceRequest - A 'FailedCreateWorkspaceRequest$WorkspaceRequest' object that contains the information about the WorkSpace that could not be created.

fcwrErrorCode - The error code.

fcwrErrorMessage - The textual error message.

fcwrWorkspaceRequest :: Lens' FailedCreateWorkspaceRequest (Maybe WorkspaceRequest)

A 'FailedCreateWorkspaceRequest$WorkspaceRequest' object that contains the information about the WorkSpace that could not be created.

fcwrErrorCode :: Lens' FailedCreateWorkspaceRequest (Maybe Text)

The error code.

fcwrErrorMessage :: Lens' FailedCreateWorkspaceRequest (Maybe Text)

The textual error message.

FailedWorkspaceChangeRequest

data FailedWorkspaceChangeRequest

Contains information about a WorkSpace that could not be rebooted (RebootWorkspaces ), rebuilt (RebuildWorkspaces ), terminated (TerminateWorkspaces ), started (StartWorkspaces ), or stopped (StopWorkspaces ).

See: failedWorkspaceChangeRequest smart constructor.

failedWorkspaceChangeRequest :: FailedWorkspaceChangeRequest

Creates a value of FailedWorkspaceChangeRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fwcrErrorCode - The error code.

fwcrWorkspaceId - The identifier of the WorkSpace.

fwcrErrorMessage - The textual error message.

fwcrErrorCode :: Lens' FailedWorkspaceChangeRequest (Maybe Text)

The error code.

fwcrWorkspaceId :: Lens' FailedWorkspaceChangeRequest (Maybe Text)

The identifier of the WorkSpace.

fwcrErrorMessage :: Lens' FailedWorkspaceChangeRequest (Maybe Text)

The textual error message.

RebootRequest

data RebootRequest

Contains information used with the RebootWorkspaces operation to reboot a WorkSpace.

See: rebootRequest smart constructor.

rebootRequest

Arguments

:: Text
rWorkspaceId

-> RebootRequest

Creates a value of RebootRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rWorkspaceId - The identifier of the WorkSpace to reboot.

rWorkspaceId :: Lens' RebootRequest Text

The identifier of the WorkSpace to reboot.

RebuildRequest

data RebuildRequest

Contains information used with the RebuildWorkspaces operation to rebuild a WorkSpace.

See: rebuildRequest smart constructor.

rebuildRequest

Arguments

:: Text
rrWorkspaceId

-> RebuildRequest

Creates a value of RebuildRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrWorkspaceId - The identifier of the WorkSpace to rebuild.

rrWorkspaceId :: Lens' RebuildRequest Text

The identifier of the WorkSpace to rebuild.

StartRequest

data StartRequest

Describes the start request.

See: startRequest smart constructor.

startRequest :: StartRequest

Creates a value of StartRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sWorkspaceId - The ID of the WorkSpace.

sWorkspaceId :: Lens' StartRequest (Maybe Text)

The ID of the WorkSpace.

StopRequest

data StopRequest

Describes the stop request.

See: stopRequest smart constructor.

stopRequest :: StopRequest

Creates a value of StopRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srWorkspaceId - The ID of the WorkSpace.

srWorkspaceId :: Lens' StopRequest (Maybe Text)

The ID of the WorkSpace.

Tag

data Tag

Describes the tag of the WorkSpace.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value of the tag.

tagKey - The key of the tag.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag.

tagKey :: Lens' Tag Text

The key of the tag.

TerminateRequest

data TerminateRequest

Contains information used with the TerminateWorkspaces operation to terminate a WorkSpace.

See: terminateRequest smart constructor.

terminateRequest

Arguments

:: Text
trWorkspaceId

-> TerminateRequest

Creates a value of TerminateRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trWorkspaceId - The identifier of the WorkSpace to terminate.

trWorkspaceId :: Lens' TerminateRequest Text

The identifier of the WorkSpace to terminate.

UserStorage

data UserStorage

Contains information about the user storage for a WorkSpace bundle.

See: userStorage smart constructor.

userStorage :: UserStorage

Creates a value of UserStorage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usCapacity - The amount of user storage for the bundle.

usCapacity :: Lens' UserStorage (Maybe Text)

The amount of user storage for the bundle.

Workspace

data Workspace

Contains information about a WorkSpace.

See: workspace smart constructor.

workspace :: Workspace

Creates a value of Workspace with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wDirectoryId - The identifier of the AWS Directory Service directory that the WorkSpace belongs to.

wState - The operational state of the WorkSpace.

wIPAddress - The IP address of the WorkSpace.

wUserName - The user that the WorkSpace is assigned to.

wSubnetId - The identifier of the subnet that the WorkSpace is in.

wBundleId - The identifier of the bundle that the WorkSpace was created from.

wWorkspaceProperties - Undocumented member.

wRootVolumeEncryptionEnabled - Specifies whether the data stored on the root volume, or C: drive, is encrypted.

wErrorCode - If the WorkSpace could not be created, this contains the error code.

wVolumeEncryptionKey - The KMS key used to encrypt data stored on your WorkSpace.

wComputerName - The name of the WorkSpace as seen by the operating system.

wWorkspaceId - The identifier of the WorkSpace.

wUserVolumeEncryptionEnabled - Specifies whether the data stored on the user volume, or D: drive, is encrypted.

wErrorMessage - If the WorkSpace could not be created, this contains a textual error message that describes the failure.

wDirectoryId :: Lens' Workspace (Maybe Text)

The identifier of the AWS Directory Service directory that the WorkSpace belongs to.

wState :: Lens' Workspace (Maybe WorkspaceState)

The operational state of the WorkSpace.

wIPAddress :: Lens' Workspace (Maybe Text)

The IP address of the WorkSpace.

wUserName :: Lens' Workspace (Maybe Text)

The user that the WorkSpace is assigned to.

wSubnetId :: Lens' Workspace (Maybe Text)

The identifier of the subnet that the WorkSpace is in.

wBundleId :: Lens' Workspace (Maybe Text)

The identifier of the bundle that the WorkSpace was created from.

wWorkspaceProperties :: Lens' Workspace (Maybe WorkspaceProperties)

Undocumented member.

wRootVolumeEncryptionEnabled :: Lens' Workspace (Maybe Bool)

Specifies whether the data stored on the root volume, or C: drive, is encrypted.

wErrorCode :: Lens' Workspace (Maybe Text)

If the WorkSpace could not be created, this contains the error code.

wVolumeEncryptionKey :: Lens' Workspace (Maybe Text)

The KMS key used to encrypt data stored on your WorkSpace.

wComputerName :: Lens' Workspace (Maybe Text)

The name of the WorkSpace as seen by the operating system.

wWorkspaceId :: Lens' Workspace (Maybe Text)

The identifier of the WorkSpace.

wUserVolumeEncryptionEnabled :: Lens' Workspace (Maybe Bool)

Specifies whether the data stored on the user volume, or D: drive, is encrypted.

wErrorMessage :: Lens' Workspace (Maybe Text)

If the WorkSpace could not be created, this contains a textual error message that describes the failure.

WorkspaceBundle

data WorkspaceBundle

Contains information about a WorkSpace bundle.

See: workspaceBundle smart constructor.

workspaceBundle :: WorkspaceBundle

Creates a value of WorkspaceBundle with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wbBundleId - The bundle identifier.

wbOwner - The owner of the bundle. This contains the owner's account identifier, or AMAZON if the bundle is provided by AWS.

wbName - The name of the bundle.

wbComputeType - A ComputeType object that specifies the compute type for the bundle.

wbUserStorage - A UserStorage object that specifies the amount of user storage that the bundle contains.

wbDescription - The bundle description.

wbBundleId :: Lens' WorkspaceBundle (Maybe Text)

The bundle identifier.

wbOwner :: Lens' WorkspaceBundle (Maybe Text)

The owner of the bundle. This contains the owner's account identifier, or AMAZON if the bundle is provided by AWS.

wbName :: Lens' WorkspaceBundle (Maybe Text)

The name of the bundle.

wbComputeType :: Lens' WorkspaceBundle (Maybe ComputeType)

A ComputeType object that specifies the compute type for the bundle.

wbUserStorage :: Lens' WorkspaceBundle (Maybe UserStorage)

A UserStorage object that specifies the amount of user storage that the bundle contains.

wbDescription :: Lens' WorkspaceBundle (Maybe Text)

The bundle description.

WorkspaceConnectionStatus

data WorkspaceConnectionStatus

Describes the connection status of a WorkSpace.

See: workspaceConnectionStatus smart constructor.

workspaceConnectionStatus :: WorkspaceConnectionStatus

Creates a value of WorkspaceConnectionStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wcsLastKnownUserConnectionTimestamp - The timestamp of the last known user connection.

wcsConnectionStateCheckTimestamp - The timestamp of the connection state check.

wcsWorkspaceId - The ID of the WorkSpace.

wcsConnectionState - The connection state of the WorkSpace. Returns UNKOWN if the WorkSpace is in a Stopped state.

wcsLastKnownUserConnectionTimestamp :: Lens' WorkspaceConnectionStatus (Maybe UTCTime)

The timestamp of the last known user connection.

wcsConnectionStateCheckTimestamp :: Lens' WorkspaceConnectionStatus (Maybe UTCTime)

The timestamp of the connection state check.

wcsWorkspaceId :: Lens' WorkspaceConnectionStatus (Maybe Text)

The ID of the WorkSpace.

wcsConnectionState :: Lens' WorkspaceConnectionStatus (Maybe ConnectionState)

The connection state of the WorkSpace. Returns UNKOWN if the WorkSpace is in a Stopped state.

WorkspaceDirectory

data WorkspaceDirectory

Contains information about an AWS Directory Service directory for use with Amazon WorkSpaces.

See: workspaceDirectory smart constructor.

workspaceDirectory :: WorkspaceDirectory

Creates a value of WorkspaceDirectory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wdRegistrationCode - The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

wdIAMRoleId - The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

wdDirectoryId - The directory identifier.

wdState - The state of the directory's registration with Amazon WorkSpaces

wdCustomerUserName - The user name for the service account.

wdSubnetIds - An array of strings that contains the identifiers of the subnets used with the directory.

wdAlias - The directory alias.

wdWorkspaceSecurityGroupId - The identifier of the security group that is assigned to new WorkSpaces.

wdDirectoryType - The directory type.

wdWorkspaceCreationProperties - A structure that specifies the default creation properties for all WorkSpaces in the directory.

wdDNSIPAddresses - An array of strings that contains the IP addresses of the DNS servers for the directory.

wdDirectoryName - The name of the directory.

wdRegistrationCode :: Lens' WorkspaceDirectory (Maybe Text)

The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

wdIAMRoleId :: Lens' WorkspaceDirectory (Maybe Text)

The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

wdDirectoryId :: Lens' WorkspaceDirectory (Maybe Text)

The directory identifier.

wdState :: Lens' WorkspaceDirectory (Maybe WorkspaceDirectoryState)

The state of the directory's registration with Amazon WorkSpaces

wdCustomerUserName :: Lens' WorkspaceDirectory (Maybe Text)

The user name for the service account.

wdSubnetIds :: Lens' WorkspaceDirectory [Text]

An array of strings that contains the identifiers of the subnets used with the directory.

wdAlias :: Lens' WorkspaceDirectory (Maybe Text)

The directory alias.

wdWorkspaceSecurityGroupId :: Lens' WorkspaceDirectory (Maybe Text)

The identifier of the security group that is assigned to new WorkSpaces.

wdDirectoryType :: Lens' WorkspaceDirectory (Maybe WorkspaceDirectoryType)

The directory type.

wdWorkspaceCreationProperties :: Lens' WorkspaceDirectory (Maybe DefaultWorkspaceCreationProperties)

A structure that specifies the default creation properties for all WorkSpaces in the directory.

wdDNSIPAddresses :: Lens' WorkspaceDirectory [Text]

An array of strings that contains the IP addresses of the DNS servers for the directory.

wdDirectoryName :: Lens' WorkspaceDirectory (Maybe Text)

The name of the directory.

WorkspaceProperties

data WorkspaceProperties

Describes the properties of a WorkSpace.

See: workspaceProperties smart constructor.

workspaceProperties :: WorkspaceProperties

Creates a value of WorkspaceProperties with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wpRunningMode - The running mode of the WorkSpace. AlwaysOn WorkSpaces are billed monthly. AutoStop WorkSpaces are billed by the hour and stopped when no longer being used in order to save on costs.

wpRunningModeAutoStopTimeoutInMinutes - The time after a user logs off when WorkSpaces are automatically stopped. Configured in 60 minute intervals.

wpRunningMode :: Lens' WorkspaceProperties (Maybe RunningMode)

The running mode of the WorkSpace. AlwaysOn WorkSpaces are billed monthly. AutoStop WorkSpaces are billed by the hour and stopped when no longer being used in order to save on costs.

wpRunningModeAutoStopTimeoutInMinutes :: Lens' WorkspaceProperties (Maybe Int)

The time after a user logs off when WorkSpaces are automatically stopped. Configured in 60 minute intervals.

WorkspaceRequest

data WorkspaceRequest

Contains information about a WorkSpace creation request.

See: workspaceRequest smart constructor.

workspaceRequest

Arguments

:: Text
wrDirectoryId

-> Text
wrUserName

-> Text
wrBundleId

-> WorkspaceRequest

Creates a value of WorkspaceRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wrWorkspaceProperties - Undocumented member.

wrRootVolumeEncryptionEnabled - Specifies whether the data stored on the root volume, or C: drive, is encrypted.

wrVolumeEncryptionKey - The KMS key used to encrypt data stored on your WorkSpace.

wrUserVolumeEncryptionEnabled - Specifies whether the data stored on the user volume, or D: drive, is encrypted.

wrTags - The tags of the WorkSpace request.

wrDirectoryId - The identifier of the AWS Directory Service directory to create the WorkSpace in. You can use the DescribeWorkspaceDirectories operation to obtain a list of the directories that are available.

wrUserName - The username that the WorkSpace is assigned to. This username must exist in the AWS Directory Service directory specified by the DirectoryId member.

wrBundleId - The identifier of the bundle to create the WorkSpace from. You can use the DescribeWorkspaceBundles operation to obtain a list of the bundles that are available.

wrWorkspaceProperties :: Lens' WorkspaceRequest (Maybe WorkspaceProperties)

Undocumented member.

wrRootVolumeEncryptionEnabled :: Lens' WorkspaceRequest (Maybe Bool)

Specifies whether the data stored on the root volume, or C: drive, is encrypted.

wrVolumeEncryptionKey :: Lens' WorkspaceRequest (Maybe Text)

The KMS key used to encrypt data stored on your WorkSpace.

wrUserVolumeEncryptionEnabled :: Lens' WorkspaceRequest (Maybe Bool)

Specifies whether the data stored on the user volume, or D: drive, is encrypted.

wrTags :: Lens' WorkspaceRequest [Tag]

The tags of the WorkSpace request.

wrDirectoryId :: Lens' WorkspaceRequest Text

The identifier of the AWS Directory Service directory to create the WorkSpace in. You can use the DescribeWorkspaceDirectories operation to obtain a list of the directories that are available.

wrUserName :: Lens' WorkspaceRequest Text

The username that the WorkSpace is assigned to. This username must exist in the AWS Directory Service directory specified by the DirectoryId member.

wrBundleId :: Lens' WorkspaceRequest Text

The identifier of the bundle to create the WorkSpace from. You can use the DescribeWorkspaceBundles operation to obtain a list of the bundles that are available.