Network.AWS.WorkSpaces.DescribeWorkspaceDirectories

Description

Retrieves information about the AWS Directory Service directories in the region that are registered with Amazon WorkSpaces and are available to your account.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

This operation returns paginated results.

Creating a Request

describeWorkspaceDirectories :: DescribeWorkspaceDirectories

Creates a value of DescribeWorkspaceDirectories with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwdNextToken - The NextToken value from a previous call to this operation. Pass null if this is the first call.

dwdDirectoryIds - An array of strings that contains the directory identifiers to retrieve information for. If this member is null, all directories are retrieved.

data DescribeWorkspaceDirectories

Contains the inputs for the DescribeWorkspaceDirectories operation.

See: describeWorkspaceDirectories smart constructor.

Request Lenses

dwdNextToken :: Lens' DescribeWorkspaceDirectories (Maybe Text)

The NextToken value from a previous call to this operation. Pass null if this is the first call.

dwdDirectoryIds :: Lens' DescribeWorkspaceDirectories (Maybe (NonEmpty Text))

An array of strings that contains the directory identifiers to retrieve information for. If this member is null, all directories are retrieved.

Destructuring the Response

describeWorkspaceDirectoriesResponse

Arguments

:: Int
dwdrsResponseStatus

-> DescribeWorkspaceDirectoriesResponse

Creates a value of DescribeWorkspaceDirectoriesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwdrsDirectories - An array of structures that contain information about the directories.

dwdrsNextToken - If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to this operation to retrieve the next set of items. This token is valid for one day and must be used within that time frame.

dwdrsResponseStatus - -- | The response status code.

data DescribeWorkspaceDirectoriesResponse

Contains the results of the DescribeWorkspaceDirectories operation.

See: describeWorkspaceDirectoriesResponse smart constructor.

Response Lenses

dwdrsDirectories :: Lens' DescribeWorkspaceDirectoriesResponse [WorkspaceDirectory]

An array of structures that contain information about the directories.

dwdrsNextToken :: Lens' DescribeWorkspaceDirectoriesResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to this operation to retrieve the next set of items. This token is valid for one day and must be used within that time frame.

dwdrsResponseStatus :: Lens' DescribeWorkspaceDirectoriesResponse Int

- | The response status code.