Network.AWS.WorkSpaces.DescribeWorkspaces

Description

Obtains information about the specified WorkSpaces.

Only one of the filter parameters, such as BundleId , DirectoryId , or WorkspaceIds , can be specified at a time.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the NextToken response member contains a token that you pass in the next call to this operation to retrieve the next set of items.

This operation returns paginated results.

Creating a Request

describeWorkspaces :: DescribeWorkspaces

Creates a value of DescribeWorkspaces with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwDirectoryId - Specifies the directory identifier to which to limit the WorkSpaces. Optionally, you can specify a specific directory user with the UserName parameter. This parameter cannot be combined with any other filter parameter.

dwWorkspaceIds - An array of strings that contain the identifiers of the WorkSpaces for which to retrieve information. This parameter cannot be combined with any other filter parameter. Because the CreateWorkspaces operation is asynchronous, the identifier it returns is not immediately available. If you immediately call DescribeWorkspaces with this identifier, no information is returned.

dwUserName - Used with the DirectoryId parameter to specify the directory user for whom to obtain the WorkSpace.

dwBundleId - The identifier of a bundle to obtain the WorkSpaces for. All WorkSpaces that are created from this bundle will be retrieved. This parameter cannot be combined with any other filter parameter.

dwNextToken - The NextToken value from a previous call to this operation. Pass null if this is the first call.

dwLimit - The maximum number of items to return.

data DescribeWorkspaces

Contains the inputs for the DescribeWorkspaces operation.

See: describeWorkspaces smart constructor.

Request Lenses

dwDirectoryId :: Lens' DescribeWorkspaces (Maybe Text)

Specifies the directory identifier to which to limit the WorkSpaces. Optionally, you can specify a specific directory user with the UserName parameter. This parameter cannot be combined with any other filter parameter.

dwWorkspaceIds :: Lens' DescribeWorkspaces (Maybe (NonEmpty Text))

An array of strings that contain the identifiers of the WorkSpaces for which to retrieve information. This parameter cannot be combined with any other filter parameter. Because the CreateWorkspaces operation is asynchronous, the identifier it returns is not immediately available. If you immediately call DescribeWorkspaces with this identifier, no information is returned.

dwUserName :: Lens' DescribeWorkspaces (Maybe Text)

Used with the DirectoryId parameter to specify the directory user for whom to obtain the WorkSpace.

dwBundleId :: Lens' DescribeWorkspaces (Maybe Text)

The identifier of a bundle to obtain the WorkSpaces for. All WorkSpaces that are created from this bundle will be retrieved. This parameter cannot be combined with any other filter parameter.

dwNextToken :: Lens' DescribeWorkspaces (Maybe Text)

The NextToken value from a previous call to this operation. Pass null if this is the first call.

dwLimit :: Lens' DescribeWorkspaces (Maybe Natural)

The maximum number of items to return.

Destructuring the Response

describeWorkspacesResponse

Arguments

:: Int
dwrsResponseStatus

-> DescribeWorkspacesResponse

Creates a value of DescribeWorkspacesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwrsNextToken - If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to this operation to retrieve the next set of items. This token is valid for one day and must be used within that time frame.

dwrsWorkspaces - An array of structures that contain the information about the WorkSpaces. Because the CreateWorkspaces operation is asynchronous, some of this information may be incomplete for a newly-created WorkSpace.

dwrsResponseStatus - -- | The response status code.

data DescribeWorkspacesResponse

Contains the results for the DescribeWorkspaces operation.

See: describeWorkspacesResponse smart constructor.

Response Lenses

dwrsNextToken :: Lens' DescribeWorkspacesResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to this operation to retrieve the next set of items. This token is valid for one day and must be used within that time frame.

dwrsWorkspaces :: Lens' DescribeWorkspacesResponse [Workspace]

An array of structures that contain the information about the WorkSpaces. Because the CreateWorkspaces operation is asynchronous, some of this information may be incomplete for a newly-created WorkSpace.

dwrsResponseStatus :: Lens' DescribeWorkspacesResponse Int

- | The response status code.