Network.AWS.WorkSpaces.TerminateWorkspaces

Description

Terminates the specified WorkSpaces.

Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is not maintained and will be destroyed. If you need to archive any user data, contact Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except SUSPENDED .

Creating a Request

terminateWorkspaces

Arguments

:: NonEmpty TerminateRequest
twTerminateWorkspaceRequests

-> TerminateWorkspaces

Creates a value of TerminateWorkspaces with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

twTerminateWorkspaceRequests - An array of structures that specify the WorkSpaces to terminate.

data TerminateWorkspaces

Contains the inputs for the TerminateWorkspaces operation.

See: terminateWorkspaces smart constructor.

Request Lenses

twTerminateWorkspaceRequests :: Lens' TerminateWorkspaces (NonEmpty TerminateRequest)

An array of structures that specify the WorkSpaces to terminate.

Destructuring the Response

terminateWorkspacesResponse

Arguments

:: Int
twrsResponseStatus

-> TerminateWorkspacesResponse

Creates a value of TerminateWorkspacesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

twrsFailedRequests - An array of structures representing any WorkSpaces that could not be terminated.

twrsResponseStatus - -- | The response status code.

data TerminateWorkspacesResponse

Contains the results of the TerminateWorkspaces operation.

See: terminateWorkspacesResponse smart constructor.

Response Lenses

twrsFailedRequests :: Lens' TerminateWorkspacesResponse [FailedWorkspaceChangeRequest]

An array of structures representing any WorkSpaces that could not be terminated.

twrsResponseStatus :: Lens' TerminateWorkspacesResponse Int

- | The response status code.